refactor(transport): extract TCP transport to module
